Refactored conversion to GEOMETRIC_MEAN
